refactors login2fa logic into a reusable function
